requirements.txt
errbot==6.1.8
PyGithub==1.55
